Add files via upload
Prep to mount the reset ISO, get a full content list (posted to GitHub). Copilot parsed the event logs and listed the faults by drivers, support software and MSS apps. This way we start out knowing what got missed the last time, what we have that just didn't have a chance to be installed and what we still need to source to complete the installer.